requirements.txt
streamlit
crewai
python-dotenv
pydantic
pyyaml
openai
crewai_tools
